SOLID OCP / 객체 미용 체조 2원칙 적용
